Move error code generation to kafka module to avoid librdkafka dependency
Otherwise go_rdkafka_generr would still require librdkafka installed
separately, which is unnecessary now with librdkafka bundled.
