+++ kafka/generated_errors.go
@@ -1,6 +1,6 @@
 package kafka
-// Copyright 2016 Confluent Inc.
-// AUTOMATICALLY GENERATED BY /home/maglun/gocode/bin/go_rdkafka_generr ON 2020-02-11 17:36:19.825643015 +0100 CET m=+0.000431212 USING librdkafka 1.4.0-RC1-devel-O0
+// Copyright 2016-2020 Confluent Inc.
+// AUTOMATICALLY GENERATED ON 2020-04-08 13:17:23.310983 +0200 CEST m=+0.001902312 USING librdkafka 1.4.0
 
 /*
 #include <librdkafka/rdkafka.h>
