+++ kafka/go_rdkafka_generr/go_rdkafka_generr.go
@@ -18,38 +18,9 @@ package main
  */
 
 import (
-	"fmt"
+	"github.com/confluentinc/confluent-kafka-go/kafka"
 	"os"
-	"strings"
-	"time"
 )
-
-/*
-#cgo pkg-config: --static rdkafka
-#cgo LDFLAGS: -lrdkafka
-#include <librdkafka/rdkafka.h>
-
-static const char *errdesc_to_string (const struct rd_kafka_err_desc *ed, int idx) {
-   return ed[idx].name;
-}
-
-static const char *errdesc_to_desc (const struct rd_kafka_err_desc *ed, int idx) {
-   return ed[idx].desc;
-}
-
-*/
-import "C"
-
-func camelCase(s string) string {
-	ret := ""
-	for _, v := range strings.Split(s, "_") {
-		if len(v) == 0 {
-			continue
-		}
-		ret += strings.ToUpper((string)(v[0])) + strings.ToLower(v[1:])
-	}
-	return ret
-}
 
 func main() {
 
@@ -61,55 +32,5 @@ func main() {
 	}
 	defer f.Close()
 
-	f.WriteString("package kafka\n")
-	f.WriteString("// Copyright 2016 Confluent Inc.\n")
-	f.WriteString(fmt.Sprintf("// AUTOMATICALLY GENERATED BY %s ON %v USING librdkafka %s\n",
-		os.Args[0], time.Now(), C.GoString(C.rd_kafka_version_str())))
-
-	var errdescs *C.struct_rd_kafka_err_desc
-	var csize C.size_t
-	C.rd_kafka_get_err_descs(&errdescs, &csize)
-
-	f.WriteString(`
-/*
-#include <librdkafka/rdkafka.h>
-*/
-import "C"
-
-// ErrorCode is the integer representation of local and broker error codes
-type ErrorCode int
-
-// String returns a human readable representation of an error code
-func (c ErrorCode) String() string {
-      return C.GoString(C.rd_kafka_err2str(C.rd_kafka_resp_err_t(c)))
-}
-
-const (
-`)
-
-	for i := 0; i < int(csize); i++ {
-		orig := C.GoString(C.errdesc_to_string(errdescs, C.int(i)))
-		if len(orig) == 0 {
-			continue
-		}
-		desc := C.GoString(C.errdesc_to_desc(errdescs, C.int(i)))
-		if len(desc) == 0 {
-			continue
-		}
-
-		errname := "Err" + camelCase(orig)
-
-		// Special handling to please golint
-		// Eof -> EOF
-		// Id -> ID
-		errname = strings.Replace(errname, "Eof", "EOF", -1)
-		errname = strings.Replace(errname, "Id", "ID", -1)
-
-		f.WriteString(fmt.Sprintf("    // %s %s\n", errname, desc))
-		f.WriteString(fmt.Sprintf("    %s ErrorCode = ErrorCode(C.RD_KAFKA_RESP_ERR_%s)\n",
-			errname, orig))
-	}
-
-	f.WriteString(")\n")
-
+	kafka.WriteErrorCodes(f)
 }
